pyproject.toml
[tool.poetry]
name = "ine5670"
version = "0.1.0"
description = ""
authors = ["Pedro Binotto <pedrosantibinotto@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = "^2.7.4"
pymongo = "^4.7.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
